requirements.txt
aprslib==0.7.2
asgiref==3.7.2
Django==4.2.16
django-cors-headers==4.3.1
django-extensions==3.2.3
gunicorn==23.0.0
paho-mqtt==2.1.0
psycopg==3.1.17
psycopg-binary==3.1.17
pydot==2.0.0
pygraphviz==1.11
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
xmltodict==0.13.0
